# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Contactos",
    "version": "1.0.0",
    "description": "CRUD de contactos (agenda telefónica)"
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "summary": "Listar todos los contactos",
        "responses": {
          "200": {
            "description": "Lista de contactos",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Contact" } }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear un contacto",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Contact" }
            }
          }
        },
        "responses": {
          "201": { "description": "Contacto creado" },
          "400": { "description": "Datos inválidos" }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "summary": "Obtener un contacto por ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Contacto encontrado" },
          "404": { "description": "No encontrado" }
        }
      },
      "put": {
        "summary": "Actualizar un contacto",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Contact" }
            }
          }
        },
        "responses": {
          "200": { "description": "Contacto actualizado" },
          "400": { "description": "Datos inválidos" },
          "404": { "description": "No encontrado" }
        }
      },
      "delete": {
        "summary": "Eliminar un contacto",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "Eliminado" },
          "404": { "description": "No encontrado" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Contact": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string" }
        },
        "required": ["name", "phone", "email"]
      }
    }
  }
}